# admin/swagger.yaml
openapi: 3.0.3
info:
  title: Swagger API Course FE
  description: |-
    This is a list api of course FE
  contact:
    email: lttd0204@gmail.com
  version: 1.0.11
servers:
  - url: http://localhost:3001
...
paths:
...
  /api/users:
    post:
      tags:
        - users
      summary: Add a new user
      description: Add a new user
      operationId: addUser
      requestBody:
        description: Create a new user
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
          application/xml:
            schema:
              $ref: "#/components/schemas/User"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      responses:
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/User"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/User"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
    get:
      tags:
        - users
      summary: Get all users
      description: get all users
      operationId: getAllUser
      parameters:
        - name: limit
          in: query
          description: Limit of list users
          required: false
          example: 10
          schema:
            type: number
        - name: page
          in: query
          description: Page of list users
          required: false
          example: 1
          schema:
            type: number
        - name: search
          in: query
          description: search of list users
          required: false
          schema:
            type: string
        - name: order
          in: query
          description: sort of list users
          required: false
          example: "created asc"
          schema:
            type: string
        - name: roleId
          in: query
          description: filter of list users by role id (if many add |)
          required: false
          example: "1|2"
          schema:
            type: string
        - name: status
          in: query
          description: filter of list users by status of user (if many add |)
          required: false
          example: "1|2"
          schema:
            type: string
        - name: cityId
          in: query
          description: filter of list users by cityId of user (if many add |)
          required: false
          example: "1|2"
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
            application/xml:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
  /api/users/{userId}:
    get:
      tags:
        - users
      summary: Get details user order by ID
      description: Get details user by id of user
      operationId: getOrderById
      parameters:
        - name: userId
          in: path
          description: ID of user that needs to be fetched
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
            application/xml:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
    delete:
      tags:
        - users
      summary: Delete User by ID
      description: Try delete user by id of user
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          description: ID of the user that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/User"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/User"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
    put:
      tags:
        - users
      summary: Update an existing user
      description: Update an existing user by Id
      operationId: updateUser
      parameters:
        - name: userId
          in: path
          description: user id of user updated
          required: true
          schema:
            type: string
      requestBody:
        description: Update an existent user in the store
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
          application/xml:
            schema:
              $ref: "#/components/schemas/User"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/User"
        required: true

      responses:
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/User"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/User"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
  /api/users/delete-many:
    delete:
      tags:
        - users
      summary: Delete many users by ids
      description: Delete many users by list ids
      operationId: deleteManyUser
      requestBody:
        description: Object chứa mảng id của user cần xóa
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userIds:
                  type: array
                  items:
                    type: string
      responses:
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
            application/xml:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
...
  /api/roles/{roleId}:
    get:
      tags:
        - roles
      summary: Get details roles order by ID
      description: Get details role by id of roles
      operationId: getRoleById
      parameters:
        - name: roleId
          in: path
          description: ID of role that needs to be fetched
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Role"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Role"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
    delete:
      tags:
        - roles
      summary: Delete rOLE by ID
      description: Try delete user by id of user
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          description: ID of the user that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Role"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Role"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
    put:
      tags:
        - roles
      summary: Update an existing role
      description: Update an existing role by Id
      operationId: updateRole
      parameters:
        - name: roleId
          in: path
          description: role id of role updated
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                permissions:
                  type: array
                  format: string
                  example: []
              required:
                - name
                - permissions
          application/xml:
            schema:
              type: object
              properties:
                name:
                  type: string
                permissions:
                  type: array
                  format: string
              required:
                - name
                - permissions
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                permissions:
                  type: array
                  format: string
              required:
                - name
                - permissions

      responses:
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Role"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Role"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    User:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        middleName:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          default: string
        phoneNumber:
          type: string
        address:
          type: string
        avatar:
          type: string
        city:
          type: string
        status:
          type: number
          default: 1
          enum: [0, 1]
        userType:
          type: number
          default: 3
          enum: [1, 2, 3]
        addresses:
          type: array
          items:
            type: object
            properties:
              firstName:
                type: string
              lastName:
                type: string
              middleName:
                type: string
              phoneNumber:
                type: string
              price:
                type: number
              discount:
                type: number
              product:
                type: string
                format: uuid
      required:
        - email
        - password
      xml:
        name: User
...
    Role:
      type: object
      properties:
        name:
          type: string
        permissions:
          type: array
      required:
        - name
        - permissions
      xml:
        name: Role
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT